requirements.txt
streamlit==1.28.0
huggingface-hub==0.19.4
python-dotenv==1.0.0
requests==2.31.0 